api/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-filter==24.2
django-simple-history==3.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
